pyproject.toml
[tool.poetry]
name = "bookmark-cleaner"
version = "0.1.0"
description = ""
authors = ["hskimim <hskimim8855@gmail.com>"]
readme = "README.md"
packages = [{include = "bookmark_cleaner"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
requests = "^2.31.0"
openai = "^1.7.2"
bs4 = "^0.0.1"
langchain = "^0.1.0"
unstructured = "^0.12.0"
ipykernel = "^6.29.0"
python-dotenv = "^1.0.0"
langchain-openai = "^0.0.2.post1"
tiktoken = "^0.5.2"
numpy = "^1.26.3"
tqdm = "^4.66.1"
scikit-learn = "^1.3.2"
pypdf = "^3.17.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
